Extract title bar to its own component
